Fix warnings
Value Conversion Issue Group:
xdelta3/xdelta3-hash.h:153:20: Implicit conversion loses integer precision: 'unsigned long' to 'int'
xdelta3/xdelta3.c:1096:35: Implicit conversion loses integer precision: 'usize_t' (aka 'unsigned long') to 'int'

Semantic Issue Group:
xdelta3/xdelta3.c:3527:6: Incompatible function pointer types passing 'int (*)(xd3_stream *)' (aka 'int (*)(struct _xd3_stream *)') to parameter of type 'int (*)(xd3_stream *, uint8_t *)' (aka 'int (*)(struct _xd3_stream *, unsigned char *)')
xdelta3/xdelta3.c:3643:42: Incompatible function pointer types passing 'int (*)(xd3_stream *, uint8_t *)' (aka 'int (*)(struct _xd3_stream *, unsigned char *)') to parameter of type 'int (*)(xd3_stream *)' (aka 'int (*)(struct _xd3_stream *)')

Unused Entity Issue Group:
xdelta3/xdelta3.c:354:1: Unused function 'xd3_rlist_length'
xdelta3/xdelta3.c:479:20: Unused function 'xd3_emit_uint32_t'

+++ xdelta3/xdelta3-hash.h
@@ -150,7 +150,7 @@ xd3_size_hashtable (xd3_stream   *stream,
     }
 
   cfg->powers[look-1] = 1;
-  for (int i = look-2; i >= 0; i--)
+  for (int i = (int)look-2; i >= 0; i--)
     {
       cfg->powers[i] = cfg->powers[i+1] * xd3_hash_multiplier;
     }
